use a base class for SearchEngine and SearchHint preferences
